LayoutTests/platform/gtk/accessibility/aria-toggle-button-with-title-expected.txt

Toggle button
Button title
This tests that a toggle button properly exposes the title when there isn't a direct relation and textUnderElement is required to be used.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Role: AXRole: toggle button
PASS tbutton1.title is 'AXTitle: Toggle button'
Role: AXRole: push button
PASS button.title is 'AXTitle: Button title'
